added categories table
m-to-m between categories and webpages
m-to-m between categories and campaigns

when parsing a webpage also setting up categories for it and thus also populating the categories table

wrote service to createCategories.ts to build them for a given webpage
wrote stub service to extract categories from webpage's html. it returns dummy data for now.