advertisement status can now be toggled
added helper in lib to get scalar fields of a model
added advertisement api handler to show or update an individual advertisement
when updating it only allows scalar fields to go through and rejects everything else

advertisement list page now shows a status field to toggle advertisement status

addedc onUpdate to the hook for advertisments to support updateing an item. this method first mutates the records cache with the newAdvertismentent with false for revalidate. this udpates cache immediately. then it makes a second mutate call which actually changes on the backend and then revalidates. this way if we do multiple updates they all get the latest cached data.

added info txToast