getting score/reason from chatgpt, added logic to parse categories
fixed checking for desktophtml to check for content as desktophtml is required not null

getting scores for all campaigns and then inserting them scores
not inserting scores we already have
also getting scores only if we have a campaign which has not yet been scored

wrote service to extract categories from webpage
it does meta tags, classes, and entry-category extraction and then cleans the values

wrote prompt service to call chatgpt and ask for categories along with scores for all campaigns of the webpage. it cleans up the text before sending in and gives back score as number with reason. it uses papaparse for converting to and from csv