for end user id setting cookie and reading its value when incoming
auction has fp & cookie column
impression no longer have end user columns and has auction as required

added package for cookie and cuid2

not storing user data when creating an impression

when creating an auction, we are checking cookie value and using it to create auction.
if no value then we create one.
we send back a cookie with value from auction object.