fixes cors issues with cookie
all cors objects now use function to generate the origin which is the request origin instead of * as * does not work when credentials are set to true

cors constructor also specify credentials true, this is to allow cookies in cors env.

cookies being set are httpOnly, sameSite & secure. this is so that they work in CORS environment