building ads for best campaign when auction is generated
moved services in to appropriate subdirectories to make things clear

createAdvertisement will increment adBuildFailCount when it fails to build ads. this happens when it gets error from chatgpt. and we use this to track failures and then prevent trying after it has failed many times.

major changes to how auctions generate handler
it has an array of messages to make it easy to see what it did from the front-end
no longer calling getAdvertisementsForUrl
instead calling getBestCampaignForWebpage and then getActiveAdsWithDetailForScoredCampaign
and when no ads found it calls processWebpageForAdCreation to build the ads
checks if status is on for all before tyring to even fetch the ads
when fetching the ads then all it has to check is if the ads are on

added service to get active ads with detail for a scored campaign

fixed ide error in service which got campaigns which have not met impression cap.

added adBuildFailCount column on scoredCampaign

major change in processWebpage, it no longer builds ads for every scored campaign on the webpage it is processing. instead we rely on building the ads for the best scored campaign which is matched when the webpage gets traffic.

added service processWebpageForAdCreation which is called when the webpage is loaded and its best matched campaign has no ads present
it gets a lock before attempting to build. this prevents other page requests trying to build it at the same time. lock is also never removed so if ad building fails, it wont try again till lock expires
it also wont build it adBuildFailCount is high.