storing isBest field on scoredCampaign
auction generate api now makes call to updateBestCampaign after it gets it

added migration to add isBest to scoredCampaign table

added service which sets no campaign as best
added service which sets a particular campaign as best. it uses transaction to ensure that when we set 1 as best, we also unset the others.

scored campaigns in the UI shows which one is matched

added service which takes the if a scoredCampaign was found and calls services to either set all off or one as best.